refactor: merge mod into qpaxos: message, instance, command now are all in qpaxos
