+++ components/epaxos/src/lib.rs
@@ -4,10 +4,7 @@ extern crate quick_error;
 #[macro_use]
 mod testutil;
 
-pub mod command;
 pub mod conf;
-pub mod instance;
-pub mod message;
 pub mod qpaxos;
 pub mod replica;
 pub mod snapshot;
